Changed type from 'double' to 'int'

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
-    public double calculate(double heightMeter, int weightKilogram) {
+    public int calculate(double heightMeter, int weightKilogram) {
         double index = weightKilogram / (heightMeter * heightMeter);
-        return index;
+        return (int) index;
     }
 }
